+++ components/modal/modal.js
@@ -32,7 +32,8 @@ const MODAL_DEFAULTS = {
   autoFocus: true,
   id: null,
   frameHeight: 180,
-  frameWidth: 46
+  frameWidth: 46,
+  beforeShow: null
 };
 
 function Modal(element, settings) {
@@ -78,7 +79,7 @@ Modal.prototype = {
     }
 
     // ensure is appended to body for new dom tree
-    if (this.settings.content) {
+    if (this.settings.content || this.settings.beforeShow) {
       this.settings.trigger = this.settings.content instanceof jQuery ? this.settings.trigger : 'immediate';
       this.appendContent();
       setTimeout(() => {
@@ -117,10 +118,15 @@ Modal.prototype = {
       } else {
         this.element.find('.modal-body').append(this.settings.content);
       }
-      if (this.settings.content instanceof jQuery) {
+
+      if (this.settings.content instanceof jQuery && !this.settings.beforeShow) {
         this.settings.content.removeClass('hidden is-hidden');
         this.settings.content.show();
       }
+    }
+
+    if (this.settings.beforeShow) {
+      this.element.find('.modal-body').append($('<div class="field"><div id="modal-busyindicator" class="busy card"></div></div>'));
     }
 
     if (!isAppended) {
@@ -140,6 +146,12 @@ Modal.prototype = {
     }
 
     utils.fixSVGIcons(this.element);
+
+    if (this.settings.beforeShow) {
+      const busyIndEl = $('#modal-busyindicator');
+      busyIndEl.busyindicator({}).data('busyindicator');
+      busyIndEl.trigger('start.busyindicator');
+    }
   },
 
   reStructure() {
@@ -339,11 +351,42 @@ Modal.prototype = {
     messageArea[0].style.width = `${messageArea.width()}px`;
   },
 
+  callSource() {
+    if (typeof this.settings.beforeShow !== 'function') {
+      return;
+    }
+
+    const self = this;
+    const response = function (content) {
+      if (content === false) {
+        return false;
+      }
+
+      $('#modal-busyindicator').trigger('complete.busyindicator');
+
+      if (!(content instanceof jQuery)) {
+        content = $(content);
+      }
+
+      self.element.find('.modal-body').empty();
+      self.element.find('.modal-body').append(content);
+
+      content.show();
+
+      return true;
+    };
+
+    const callBackOpts = {};
+    this.settings.beforeShow(response, callBackOpts);
+    self.open(true);
+  },
+
   /**
-  * Open the modal via the api.
-  * @returns {void}
-  */
-  open() {
+   * *
+   * Open the modal via the api.
+   * @param {boolean} ajaxReturn Flag used internally to denote its an ajax result return.
+   */
+  open(ajaxReturn) {
     let messageArea = null;
     let elemCanOpen = true;
 
@@ -380,6 +423,14 @@ Modal.prototype = {
       this.overlay.remove();
       this.root[0].style.display = 'none';
       return;
+    }
+
+    if (!ajaxReturn) {
+      this.callSource();
+
+      if (this.settings.beforeShow) {
+        return;
+      }
     }
 
     // Look for other nested dialogs and adjust the zindex.